pyproject.toml
[tool.poetry]
name = "ingest"
version = "0.1.0"
description = ""
authors = ["ywatanabe <yu.w.tennis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-storage = "~3"
google-cloud-bigquery = "~3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pylint = "~2.15"
mypy = "^1.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
